\name{SynchWave-internal}
\alias{SynchWave-internal}
\alias{padsignal}
\alias{p2up}
\alias{padarray}
\alias{cwt_freq_direct}
\alias{cwt_freq}
\alias{diff_so}
\alias{synsq_cwt_squeeze}
\alias{imdilate}
\alias{synsq_adm}
\alias{quadgk}
\alias{cwt_adm}

\title{Internal SynchWave Functions}
\description{
  The \code{padsignal}, \code{p2u}, \code{padarray}, \code{cwt_freq_direct}, 
\code{cwt_freq}, \code{diff_so}, \code{synsq_cwt_squeeze}, \code{imdilate}, \code{synsq_adm},
\code{quadgk}, \code{cwt_adm} are internal SynchWave functions.

\code{padarray} and \code{imdilate} is adapted from Matlab.
\code{quadgk} is from pracma packake 1.4.5 authored by Hans W Borchers (hwborchers at googlemail.com).
Other codes are translated from MATLAB Synchrosqueezing Toolbox, version 1.1 developed by Eugene Brevdo (http://www.math.princeton.edu/~ebrevdo/).
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
